• Reviewed all 22 classes under src/main/java and implemented   additional fixes beyond the existing review commit.
  Key changes:

  - Fixed a close/put race in src/main/java/com/landawn/abacus/
    cache/CaffeineCache.java:202 that could retain a late entry
    after closure while keeping normal puts lock-free.

  - Fixed a per-cache executor and shutdown-hook leak in src/main/
    java/com/landawn/abacus/cache/AbstractOffHeapCache.java:1522.

  - Prevented overflow in hit/miss rate calculations in src/main/
    java/com/landawn/abacus/cache/CacheStats.java:153 and src/main/
    java/com/landawn/abacus/cache/OffHeapCacheStats.java:205.

  - Replaced the misleading Caffeine per-entry-expiration TODO with
    an accurate API explanation.

  - Added regression tests for all fixes.

+++ src/main/java/com/landawn/abacus/cache/CacheStats.java
@@ -151,8 +151,10 @@ public record CacheStats(int capacity, int size, long putCount, long getCount, l
      * @see #missRate()
      */
     public double hitRate() {
-        final long requestCount = hitCount + missCount;
-        return requestCount == 0 ? 0.0 : (double) hitCount / requestCount;
+        // Convert before adding so two valid, non-negative long counters cannot overflow their
+        // denominator and produce a negative or out-of-range rate.
+        final double requestCount = (double) hitCount + missCount;
+        return requestCount == 0.0D ? 0.0D : hitCount / requestCount;
     }
 
     /**
@@ -164,7 +166,7 @@ public record CacheStats(int capacity, int size, long putCount, long getCount, l
      * @see #hitRate()
      */
     public double missRate() {
-        final long requestCount = hitCount + missCount;
-        return requestCount == 0 ? 0.0 : (double) missCount / requestCount;
+        final double requestCount = (double) hitCount + missCount;
+        return requestCount == 0.0D ? 0.0D : missCount / requestCount;
     }
 }
